realdragon 아마도 안전하게 머지

+++ Source/RuinLooters/SaveGame/RLSaveGame.h
@@ -44,10 +44,10 @@ public:
 	int32 PlayerMoney;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Player")
-	int32 CurrentHp;
+	float CurrentHp;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Player")
-	int32 MaxHp;
+	float MaxHp;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Player")
 	int32 AttackDamage;
